+++ lp-simulation-environment/monitoring/src/main/java/eu/learnpad/simulator/mon/BPMN/ModelLoader.java
@@ -1,4 +1,4 @@
-package eu.learnpad.simulator.mon.coverage;
+package eu.learnpad.simulator.mon.BPMN;
 
 import java.io.IOException;
