+++ lp-simulation-environment/monitoring/src/main/java/eu/learnpad/simulator/mon/BPMN/PathExplorer.java
@@ -1,8 +1,10 @@
-package eu.learnpad.simulator.mon.coverage;
+package eu.learnpad.simulator.mon.BPMN;
 
 import java.util.Vector;
 
 import org.w3c.dom.Document;
+
+import eu.learnpad.simulator.mon.coverage.Activity;
 
 public interface PathExplorer {
