+++ lp-simulation-environment/monitoring/src/main/java/eu/learnpad/simulator/mon/rules/RuleElements.java
@@ -1,4 +1,4 @@
-package eu.learnpad.simulator.mon.rulesGenerator;
+package eu.learnpad.simulator.mon.rules;
 
 import eu.learnpad.simulator.mon.coverage.Activity;
