+++ lp-simulation-environment/monitoring/src/main/java/eu/learnpad/simulator/mon/rules/generator/RulesPerPath.java
@@ -1,4 +1,4 @@
-package eu.learnpad.simulator.mon.rulesGenerator;
+package eu.learnpad.simulator.mon.rules.generator;
 
 import java.util.Vector;
 
@@ -8,7 +8,7 @@ import eu.learnpad.simulator.mon.coverage.Path;
 import it.cnr.isti.labse.glimpse.xml.complexEventRule.ComplexEventRuleActionListDocument;
 import it.cnr.isti.labse.glimpse.xml.complexEventRule.ComplexEventRuleType;
 
-public interface PathRulesGenerator {
+public interface RulesPerPath {
 	//ComplexEventRuleActionListDocument generateAllPathsRules(Vector<Activity[]> theUnfoldedBusinessProcess, String idBpmn);
 	Vector<Path> generateAllPaths(Vector<Activity[]> theUnfoldedBusinessProcess, String idBpmn);
 	ComplexEventRuleType generateRuleForSinglePath(Activity[] anActivitiesSet, String rulesName, String idBPMN, String idPath);
