--- lp-simulation-environment/monitoring/src/main/java/eu/learnpad/simulator/mon/rulesGenerator/PathCrossingRulesGenerator.java
@@ -1,9 +1,0 @@
-package eu.learnpad.simulator.mon.rulesGenerator;
-
-import java.util.List;
-
-import it.cnr.isti.labse.glimpse.xml.complexEventRule.ComplexEventRuleActionListDocument;
-
-public interface PathCrossingRulesGenerator {
-	ComplexEventRuleActionListDocument generateRules(List<String[]> theUnfoldedBusinessProcess);
-}
